% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{custom_colored_layout}
\alias{custom_colored_layout}
\title{Custom colored layout for console output}
\usage{
custom_colored_layout(
  level,
  msg,
  namespace = NA_character_,
  .logcall = sys.call(),
  .topcall = sys.call(-1),
  .topenv = parent.frame(),
  .timestamp = Sys.time()
)
}
\arguments{
\item{level}{Log level object}

\item{msg}{Log message}

\item{namespace}{The namespace (package) generating the log (defaults to "global")}

\item{.logcall}{The logging call being evaluated}

\item{.topcall}{R expression from which the logging function was called}

\item{.topenv}{Original frame of the .topcall calling function}

\item{.timestamp}{The timestamp when logging occurred}
}
\value{
Formatted log string with ANSI color codes
}
\description{
Creates a colored log layout with the format:
"{time} {level} [R] {namespace} : {msg}"
where level is displayed in bold color.
}
