% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R, R/utilities.R
\name{enable_windows_colors}
\alias{enable_windows_colors}
\title{Enable colors in Windows}
\usage{
enable_windows_colors()

enable_windows_colors()
}
\value{
NULL invisibly
}
\description{
Sets the crayon.enabled option to TRUE when running on Windows,
which enables ANSI color support in R console on Windows.
}
