% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R, R/utilities.R
\name{get_common_log_file}
\alias{get_common_log_file}
\title{Get the common log file path}
\usage{
get_common_log_file()

get_common_log_file()
}
\value{
Character string with the full path to the log file
}
\description{
Returns a standardized path for the common log file based on
environment variables or default locations. Creates the
parent directory if it doesn't exist.
}
