% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{update_log_level}
\alias{update_log_level}
\title{Update log threshold for a namespace}
\usage{
update_log_level(namespace, level, console_only = FALSE, file_only = FALSE)
}
\arguments{
\item{namespace}{The namespace to update}

\item{level}{New log level threshold}

\item{console_only}{Update only console threshold}

\item{file_only}{Update only file threshold}
}
\description{
Update log threshold for a namespace
}
